+++ DistroLauncher/AfterInstall.h
@@ -1,0 +1,31 @@
+/*
+ * Copyright (C) 2022 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+#pragma once
+
+#include <sstream>
+
+#include "DistributionInfo.h"
+
+// Replace with std::wstring_view::starts_with and ends_with in C++20
+[[nodiscard]] bool starts_with(const std::wstring_view str, const std::wstring_view other);
+[[nodiscard]] bool ends_with(const std::wstring_view str, const std::wstring_view other);
+
+[[nodiscard]] std::wstring_view GetDefaultUpgradePolicy();
+
+HRESULT OverrideUpgradePolicy(DWORD& exitCode);
+
+HRESULT AfterInstall(DWORD& exitCode);
